Performance fix: avoid redundant analyses.
Flake8 is relatively fast, but do_update_state is called many time
at each load and save, making the whole thing slow.
(1 second to save on a SSD-powered laptop?!)

According to the docstring, the method is only there to start one analysis
on load. The 'loaded' signal does a better job at this (it is only triggered
once per call) and works the same as 'saved'.

Loading and saving (on my machine) is still noticably slower than without
the plugin, but it is now much better than before.